pythonfuncsqltest/requirements.txt
bs4
pandas
pysftp
azure-functions
urllib3
requests
sqlalchemy
pyodbc
numpy
datetime